+++ ingredient/ingredient-event-hub-namespace/src/plugin.ts
@@ -1,7 +1,9 @@
-import { BaseIngredient, IngredientManager } from "@azbake/core"
+import { BaseIngredient, IngredientManager, BakeVariable } from "@azbake/core"
 import { ARMHelper } from "@azbake/arm-helper"
 import ARMTemplate from "./arm.json"
+import stockAlerts from "./stockAlerts.json"
 import { EventHubNamespaceUtils } from "./functions.js";
+import { stringify } from "querystring";
 
 export class EventHubNamespacePlugin extends BaseIngredient {
     public async Execute(): Promise<void> {
@@ -10,7 +12,7 @@ export class EventHubNamespacePlugin extends BaseIngredient {
             this._logger.log('Event Hub Namespace Plugin Logging: ')
 
             const helper = new ARMHelper(this._ctx);
-
+                                
             let params = await helper.BakeParamsToARMParamsAsync(this._name, this._ingredient.properties.parameters)
 
             if (!params["diagnosticsEnabled"])
@@ -30,6 +32,10 @@ export class EventHubNamespacePlugin extends BaseIngredient {
             }
 
             await helper.DeployTemplate(this._name, ARMTemplate, params, await util.resource_group())
+            
+            let alertTarget = params["eventHubNamespaceName"].value
+            let alertOverrides = this._ingredient.properties.alerts
+            await helper.DeployAlerts(this._name, await util.resource_group(), alertTarget, stockAlerts, alertOverrides)
 
         } catch (error) {
             this._logger.error('deployment failed: ' + error)